Newly created tanks_fixtures.py and modified modules conftest and liquid_fixtures.
